backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
supabase==2.11.0
httpx>=0.24,<0.28
websockets>=13.0
python-multipart==0.0.6
pillow>=12.1.0
google-generativeai==0.3.2
python-dotenv==1.0.0
